programm to startup

+++ Components/Camera.cs
@@ -4,7 +4,12 @@ namespace ConsoleEngine.Components
 {
     public class Camera : Component
     {
-        public double[] Offset = {GameConfig.Data.WIDTH/2.0/Symbol.FiveByEight, GameConfig.Data.HEIGHT/2.0};
+        public double[] Offset =
+        {
+            (int) (GameConfig.Data.WIDTH / 2.0 / Symbol.FiveByEight),
+            (int) (GameConfig.Data.HEIGHT / 2.0)
+        };
+
         public double a;
         public bool b;
         public string c;
